feat: Modular component refactor and UI improvements
- Refactored monolithic App.tsx (497→258 lines) to use modular components
- Created Header component with glassmorphism design and tone control
- Created Sidebar component with Content History and backend status
- Created PostCard component with modern card design and media handling
- Added shared TypeScript interfaces in types/shared.ts to eliminate redundancy
- Implemented complete API integration for rephraseContent and approveAndPost
- Updated CSS with modern glassmorphism styling and improved animations
- Fixed sidebar positioning from middle-left to top-left layout
- Enhanced responsive design and mobile overlay functionality
- Eliminated duplicate code and improved component separation

Technical improvements:
- Centralized type definitions for Author, Post, and PushHistory interfaces
- Proper error handling and fallback for MCP integration
- Improved state management across components
- Better UX with loading states and notifications
- Maintained all existing functionality while improving code organization
